+++ src/Prime/Storage/Filter/Criteria/DelayCriteria.php
@@ -46,14 +46,9 @@ final readonly class DelayCriteria implements CriteriaInterface
 
         $delay = $stamp->getDelay();
 
-        if (null === $this->maxDelay) {
-            return $delay >= $this->minDelay;
-        }
+        $meetsMin = null === $this->minDelay || $delay >= $this->minDelay;
+        $meetsMax = null === $this->maxDelay || $delay <= $this->maxDelay;
 
-        if ($delay <= $this->maxDelay) {
-            return $delay >= $this->minDelay;
-        }
-
-        return false;
+        return $meetsMin && $meetsMax;
     }
 }